pyproject.toml
[tool.poetry]
package-mode = false
name = "prada-id"
version = "0.1.0"
description = "A service for classifying vintage Prada clothing by season"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
pillow = "^10.2.0"
numpy = "^1.26.4"
scikit-learn = "^1.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
torch = "^2.6.0"
opencv-python = "^4.11.0.86"
torchvision = "^0.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 